Use subdomains for event slugs
Switch to using subdomains (as opposed to URL paths) for event slugs.
This currently has feature parity with the old approach. Still to do:

- [x] Serve Flask-Security pages at the subdomain
- [x] Serve the admin at the subdomain
- [ ] Use the slug to filter content in the admin
- [x] Deduplicate event detection logic between the admin and frontend
- [ ] Serve static assets at the subdomain

Notes:

- This change requires setting the application's server name during
  development.
- Remaining todo items are nice to have but not required.
