+++ pygotham/factory.py
@@ -1,10 +1,12 @@
 """Application factory."""
 
-from flask import Flask
+import arrow
+from flask import Flask, g
 from flask_security import SQLAlchemyUserDatastore
+from sqlalchemy import or_
 
 from pygotham.core import copilot, db, mail, migrate, security
-from pygotham.models import Role, User
+from pygotham.models import Event, Role, User
 from pygotham.utils import check_required_settings, register_blueprints
 
 __all__ = ('create_app',)
@@ -40,5 +42,31 @@ def create_app(package_name, package_path, settings_override=None,
     )
 
     register_blueprints(app, package_name, package_path)
+
+    @app.url_defaults
+    def add_event_slug(endpoint, values):
+        if 'event_slug' in values or not g.current_event:
+            return
+        if app.url_map.is_endpoint_expecting(endpoint, 'event_slug'):
+            values['event_slug'] = g.current_event.slug
+
+    @app.url_value_preprocessor
+    def current_event_from_url(endpoint, values):
+        if values is None:
+            values = {}
+        if endpoint and app.url_map.is_endpoint_expecting(endpoint, 'event_slug'):
+            now = arrow.utcnow().to('America/New_York').naive
+            g.current_event = Event.query.filter(
+                Event.slug == values.pop('event_slug', None),
+                Event.active == True,
+                or_(Event.activity_begins == None, Event.activity_begins <= now),
+                or_(Event.activity_ends == None, Event.activity_ends > now),
+            ).order_by(Event.activity_begins).first_or_404()
+        else:
+            g.current_event = Event.query.first()
+
+    @app.context_processor
+    def current_event():
+        return {'current_event': getattr(g, 'current_event', None)}
 
     return app
