+++ pygotham/settings.py
@@ -12,6 +12,7 @@ def bool_(key, default):
 
 DEBUG = bool_('DEBUG', False)
 SECRET_KEY = env.get('SECRET_KEY', DOES_NOT_EXIST)
+SERVER_NAME = env.get('SERVER_NAME')
 
 GOOGLE_ANALYTICS_PROFILE_ID = env.get('GOOGLE_ANALYTICS_PROFILE_ID')
 
@@ -44,6 +45,7 @@ SECURITY_POST_REGISTER_VIEW = 'profile.settings'
 SECURITY_RECOVERABLE = True
 SECURITY_REGISTERABLE = True
 SECURITY_SEND_REGISTER_EMAIL = bool_('SECURITY_SEND_REGISTER_EMAIL', True)
+SECURITY_SUBDOMAIN = '<event_slug>'
 SECURITY_TRACKABLE = True
 
 # Flask-SQLAlchemy
